Add recruiter layout
Recruiter has home page with navbar at the top
