+++ frontend/src/app/app.component.html
@@ -1,2 +1,2 @@
 <!-- <app-menu></app-menu> -->
-<router-outlet></router-outlet>
+<router-outlet></router-outlet>`
