--- frontend/src/app/dashboard/dashboard.module.spec.ts
@@ -1,13 +1,0 @@
-import { DashboardModule } from './dashboard.module';
-
-describe('DashboardModule', () => {
-  let dashboardModule: DashboardModule;
-
-  beforeEach(() => {
-    dashboardModule = new DashboardModule();
-  });
-
-  it('should create an instance', () => {
-    expect(dashboardModule).toBeTruthy();
-  });
-});
