--- frontend/src/app/home-page/home-page.component.css
@@ -1,17 +1,0 @@
-.btn {
-    margin: 15vh 0 15vh 0;
-}
-
-.dataBackground {
-    margin: auto;
-    margin-bottom: 5vh;
-}
-
-.buttonSpan {
-    text-align: center;
-}
-
-h1 {
-    margin: auto;
-    margin-bottom: 10vw;
-}
