--- frontend/src/app/home-page/home-page.component.html
@@ -1,13 +1,0 @@
-<div class="container col-lg-10">
-    <div class="row header">
-        <h1 class="col-lg-6">Witaj w aplikacji do przeprowadzania testów rekrutacyjnych</h1>
-    </div>
-    <!-- <div class="dataBackground col-lg-4 row rounded">
-        <span class="col-lg-12 buttonSpan">
-            <button type="button" class="btn btn-success btn-lg btn-login col-lg-12" [routerLink]="['/login']">Zaloguj się</button>
-        </span>
-         <span class="col-lg-6 buttonSpan">
-            <button type="button" class="btn btn-success btn-lg btn-register col-lg-8" [routerLink]="['/register']">Zarejestruj się</button>
-        </span>
-    </div> -->
-</div>
