--- frontend/src/app/home-page/home-page.component.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-home-page',
-  templateUrl: './home-page.component.html',
-  styleUrls: ['./home-page.component.css']
-})
-export class HomePageComponent implements OnInit {
-
-  constructor() { console.log('home component');}
-
-  ngOnInit() {
-    console.log('home ng component');
-  }
-
-}
