--- frontend/src/app/navbar/navbar.component.html
@@ -1,1 +1,0 @@
-<p>navbar works!</p>
