+++ frontend/src/app/recruiter/layout/layout.component.html
@@ -1,0 +1,1 @@
+<recruiter-navbar></recruiter-navbar>
