+++ frontend/src/app/recruiter/layout/navbar/menu/menu.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { AuthenticationService } from '../shared/authentication.service';
+import { AuthenticationService } from '../../../../shared/authentication.service';
+import {MatMenuModule} from '@angular/material/menu'; 
 
 @Component({
   selector: 'app-menu',
