+++ frontend/src/app/recruiter/layout/navbar/navbar.component.html
@@ -1,0 +1,7 @@
+<mat-toolbar>
+    <mat-toolbar-row>
+      <app-menu></app-menu>
+      <button id='logoutButton' mat-raised-button (click)="logout()">Logout</button>
+    </mat-toolbar-row>
+</mat-toolbar>
+
