+++ frontend/src/app/recruiter/navbar/menu/menu.component.ts
@@ -1,0 +1,17 @@
+import { Component, OnInit } from '@angular/core';
+import { AuthenticationService } from '../../../shared/authentication.service';
+import {MatMenuModule} from '@angular/material/menu'; 
+
+@Component({
+  selector: 'app-menu',
+  templateUrl: './menu.component.html',
+  styleUrls: ['./menu.component.css']
+})
+export class MenuComponent implements OnInit {
+
+  constructor(private auth: AuthenticationService) { }
+
+  ngOnInit() {
+  }
+
+}
