+++ frontend/src/app/recruiter/sample/sample.component.html
@@ -1,1 +1,3 @@
+<app-menu></app-menu>
+<recruiter-navbar></recruiter-navbar>
 <p>sample works!</p>
