+++ frontend/src/app/recruiter/users-list/users-list.component.ts
@@ -1,9 +1,9 @@
 import {Component, Inject, OnInit} from '@angular/core';
-import {AuthenticationService} from '../shared/authentication.service';
+import {AuthenticationService} from '../../shared/authentication.service';
 import {Router} from '@angular/router';
-import {UsersManagementService} from '../shared/users-management.service';
+import {UsersManagementService} from '../../shared/users-management.service';
 import {MAT_DIALOG_DATA, MatDialog, MatDialogRef} from '@angular/material';
-import {ChangePasswordDialog, DialogData} from '../login/login.component';
+import {ChangePasswordDialog, DialogData} from '../../login/login.component';
 
 
 export interface DialogNewUserData {
